Resume-Parser-OpenAI-main/requirements.txt
Flask==3.0.2
pypdf==4.1.0
pyyaml==6.0
google-generativeai==0.3.1